+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/pages/level-one/level-one.component.js
@@ -10,6 +10,7 @@ var LevelOneComponent = (function () {
     }
     LevelOneComponent.prototype.ngOnInit = function () {
         this.title = config_1.Config.title + ' - About';
+        this.nextScreen = '/view/level-two/level-two';
     };
     return LevelOneComponent;
 }());