+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/shared/score/score.ts
@@ -1,0 +1,8 @@
+export class Score {
+    id: number;
+    user: string;
+    time: string;
+    moves: number;
+    level: number;
+    cssClass: string;
+}